updated readme and spel changes done
