--- CallRecoding/src/main/java/com/acsrecording/api/Controller/RestController.java
@@ -1,5 +1,0 @@
-package com.acsrecording.api.Controller;
-
-public @interface RestController {
-
-}
