--- CallRecoding/src/main/resources/config.properties
@@ -1,4 +1,0 @@
-Connectionstring=endpoint=https://acs-app-validations.communication.azure.com/;accesskey=F7/3U887zz4Ur5CZwhQmGLml8y9uWVb6ir4kqI2K4FyGtShWFxdYBLBf4ZxtJ41P6C5KUC2sZpfayCOtl9DMbw==
-ACSAcquiredPhoneNumber=+18332197422
-BaseUri=https://3b6d-103-183-203-119.ngrok.io
-targetPhoneNumber=+918374734062
