--- CallRecoding/src/test/java/com/acsrecording/api/ApiApplicationTests.java
@@ -1,20 +1,0 @@
-package com.acsrecording.api;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
-
-/**
- * Unit test for simple App.
- */
-public class ApiApplicationTests 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
