+++ CallRecording/pom.xml
@@ -1,28 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
+
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.5.1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.acscalling</groupId>
-	<artifactId>recordingapi</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>api</name>
-	<description>Apis for ACS calling</description>
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
+  <groupId>com.acscalling</groupId>
+  <artifactId>recordingapi</artifactId>
+  <version>1.0-SNAPSHOT</version>
+
+  <name>api</name>
+  <!-- FIXME change it to the project's website -->
+  <url>http://www.example.com</url>
+
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <maven.compiler.source>1.7</maven.compiler.source>
+    <maven.compiler.target>1.7</maven.compiler.target>
 		<java.version>11</java.version>
 		<node.version>v14.17.0</node.version>
 		<jib-maven-plugin.version>2.5.2</jib-maven-plugin.version>
 		<docker.image.prefix>{docker.image.prefix}</docker.image.prefix>
-	</properties>
-	<dependencies>		
+  </properties>
+
+ <dependencies>		
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
@@ -75,14 +80,8 @@
 			</snapshots>
 	</repository>
 	</repositories>
-	<!-- <repositories>
-		<repository>
-			<id>maven-cognitiveservices-speech</id>
-			<name>communication-callingserver-sdk</name>
-			<url>https://csspeechstorage.blob.core.windows.net/maven/</url>
-		</repository>
-	</repositories> -->
-	<build>
+
+  <build>
 		<pluginManagement>
 			<plugins>
 			<plugin>
