+++ CallRecording/src/main/java/com/acsrecording/api/ApiApplication.java
@@ -1,15 +1,13 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
 package com.acsrecording.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ApiApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(ApiApplication.class, args);
-	}
-
+public class ApiApplication 
+{
+    public static void main( String[] args )
+    {
+        SpringApplication.run(ApiApplication.class, args);
+    }
 }
