+++ CallRecording/src/main/java/com/acsrecording/api/ConfigurationManager.java
@@ -1,6 +1,3 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
 package com.acsrecording.api;
 
 import java.io.InputStream;
