+++ CallRecording/src/main/resources/config.properties
@@ -1,4 +1,4 @@
 Connectionstring=%Connectionstring%
 ACSAcquiredPhoneNumber=%ACSAcquiredPhoneNumber%
 BaseUri=%BaseUri%
-targetPhoneNumber=targetPhoneNumber
+targetPhoneNumber=%targetPhoneNumber%
