+++ CallRecording/src/test/java/com/acsrecording/api/ApiApplicationTests.java
@@ -1,13 +1,20 @@
 package com.acsrecording.api;
 
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class ApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
+/**
+ * Unit test for simple App.
+ */
+public class ApiApplicationTests 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
 }
